refactor: give the tunnel code boundaries
`provisioner/tunnel.rs` had grown to 1,483 lines holding four unrelated
concerns: pure block arithmetic, generic interface planning, marketplace
node allocation, and the marketplace data-plane document — including a
constant naming a Linux bridge on somebody else's hardware, sitting next
to the generic planner. The apply half was stranded in `worker.rs` among
everything else the worker does, and two public types were both called
`Tunnel`.

Split along what each part is allowed to know:

  wg/address.rs         block arithmetic          knows nothing, pure
  wg/plan.rs            desired interface state   db, pool, tunnel
  wg/apply.rs           drift and push            plan, TunnelRouter
  marketplace_tunnel.rs node data plane           nodes, hosts, guests
  vpn.rs                device registration       plans, services

`PoolPlan` becomes `InterfacePlan`, because a pool no longer decides it
alone: an interface terminating a VPN service is addressed from the
service's block.

`router::Tunnel` becomes `router::ObservedInterface`. It described what
was read off a route server while `lnvps_db::Tunnel` describes what
should be there — opposite meanings under one name, in files that import
both.

The apply half moves into a `TunnelRealiser { db }`, the same shape as
`NetworkProvisioner`, with the worker reaching it through a `tunnels()`
accessor. `worker.rs` loses 340 lines.

Pure move and rename; no behaviour change. Every test passes unchanged
apart from imports.
