+++ lnvps_api/src/provisioner/mod.rs
@@ -1,13 +1,14 @@
 mod ip_range;
+mod marketplace_tunnel;
 mod migrate;
 mod probe;
 mod probe_address;
 #[cfg(feature = "linux-ssh")]
 mod probe_ssh;
-mod tunnel;
 mod vm;
 mod vm_network;
 mod vpn;
+pub mod wg;
 
 #[cfg(test)]
 mod retry_tests;
@@ -20,12 +21,13 @@ mod rollback_tests;
 
 pub use ip_range::*;
 pub use lnvps_api_common::{HostCapacityService, NetworkProvisioner, PricingEngine};
+pub use marketplace_tunnel::*;
 pub use migrate::*;
 pub use probe::*;
 pub use probe_address::*;
 #[cfg(feature = "linux-ssh")]
 pub use probe_ssh::*;
-pub use tunnel::*;
 pub use vm::*;
 pub use vm_network::*;
 pub use vpn::*;
+pub use wg::*;
