+++ lnvps_api/src/worker.rs
@@ -136,53 +136,6 @@ struct HostInfoOutput {
     gpu_features: Vec<String>,
 }
 
-/// What a tunnel pool's route server disagreed with the database about.
-///
-/// Kept as three lists rather than a count because they mean different things:
-/// a peer that is *missing* was configured and is gone, a *changed* one is
-/// carrying the wrong anti-spoof list, and an *unclaimed* one is a key on an
-/// LNVPS interface that no allocation accounts for.
-#[derive(Debug, Clone, Default, PartialEq, Eq)]
-pub struct TunnelPeerDrift {
-    /// Allocated peers the route server did not have
-    pub missing: Vec<String>,
-    /// Peers whose allowed IPs no longer matched their allocation
-    pub changed: Vec<String>,
-    /// Peers on the interface that no tunnel claims
-    pub unclaimed: Vec<String>,
-}
-
-impl TunnelPeerDrift {
-    pub fn is_empty(&self) -> bool {
-        self.missing.is_empty() && self.changed.is_empty() && self.unclaimed.is_empty()
-    }
-}
-
-impl std::fmt::Display for TunnelPeerDrift {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(
-            f,
-            "{} missing, {} changed, {} unclaimed",
-            self.missing.len(),
-            self.changed.len(),
-            self.unclaimed.len()
-        )
-    }
-}
-
-/// Whether two peers permit the same set of addresses.
-///
-/// Compared as a set: `wg` reports allowed IPs in its own order, and treating
-/// that as a difference would rewrite a working peer's anti-spoof list on every
-/// single reconcile.
-fn same_allowed_ips(a: &crate::router::WireguardPeer, b: &crate::router::WireguardPeer) -> bool {
-    let mut x: Vec<&String> = a.allowed_ips.iter().collect();
-    let mut y: Vec<&String> = b.allowed_ips.iter().collect();
-    x.sort();
-    y.sort();
-    x == y
-}
-
 /// Primary background worker logic
 /// Handles deleting expired VMs and sending notifications
 #[derive(Clone)]
@@ -583,6 +536,15 @@ fn format_lead_window(lead: TimeDelta) -> String {
 }
 
 impl Worker {
+    /// Applying interface plans to route servers.
+    ///
+    /// A thin accessor rather than a field: the realiser holds only the
+    /// database, so constructing one per call costs an `Arc` clone and keeps
+    /// the worker from carrying a service it uses in four places.
+    fn tunnels(&self) -> crate::provisioner::wg::apply::TunnelRealiser {
+        crate::provisioner::wg::apply::TunnelRealiser::new(self.db.clone())
+    }
+
     /// Grace period (in days) for a subscription, tiered by subscription age.
     /// Newer subscriptions get shorter grace windows so resources aren't held open
     /// for days after a brand-new VM expires.
@@ -857,7 +819,7 @@ impl Worker {
                         .iter()
                         .filter(|p| p.router_id == router_id && p.enabled)
                     {
-                        if let Err(e) = self.reconcile_tunnel_peers(pool.id).await {
+                        if let Err(e) = self.tunnels().reconcile_tunnel_peers(pool.id).await {
                             warn!("Failed to reconcile tunnel pool {}: {}", pool.id, e);
                         }
                     }
@@ -959,312 +921,6 @@ impl Worker {
             .set_router_bgp_session_enabled(router_id, cache_name, enabled)
             .await
             .map_err(|e| anyhow!("failed to persist BGP session enabled flag: {}", e))?;
-        Ok(())
-    }
-
-    /// Enable/disable a tunnel on a router and refresh its cached state.
-    pub async fn toggle_tunnel(&self, router_id: u64, name: &str, enabled: bool) -> Result<()> {
-        let router = crate::router::get_router(&self.db, router_id)
-            .await
-            .map_err(|e| anyhow!("failed to load router {}: {}", router_id, e))?;
-        let tr = router.tunnel().context("router does not support tunnels")?;
-        // The admin API addresses tunnels by name (the cache key), but the backend
-        // toggles by its own id (interface name on Linux, `<kind>:<.id>` on
-        // Mikrotik). Resolve the id from the live listing.
-        let tunnels = tr
-            .list_tunnels()
-            .await
-            .map_err(|e| anyhow!("failed to list tunnels: {}", e))?;
-        let target = tunnels
-            .iter()
-            .find(|t| t.name == name)
-            .context("tunnel not found")?;
-        let id = target.id.as_deref().unwrap_or(name);
-        tr.set_tunnel_enabled(id, enabled)
-            .await
-            .map_err(|e| anyhow!("failed to toggle tunnel: {}", e))?;
-        // Refresh the cached inventory so the admin API reflects the new state.
-        // The tunnel `enabled` flag is discovery-authoritative (the interface
-        // up/down state), so re-listing after the change is sufficient.
-        if let Ok(tunnels) = tr.list_tunnels().await {
-            for t in &tunnels {
-                if let Err(e) = self.db.upsert_router_tunnel(&t.to_db(router_id)).await {
-                    warn!("Failed to refresh tunnel cache: {}", e);
-                }
-            }
-        }
-        Ok(())
-    }
-
-    /// Configure a tunnel pool's WireGuard interface on its route server.
-    ///
-    /// This is a **push**, not a reconcile: LNVPS generates and holds the
-    /// interface's key material, so what the database says is what the
-    /// interface should be. Without it a pool could only describe an interface
-    /// somebody configured by hand, and bringing up a new route server would be
-    /// a manual job with a database row bolted on afterwards.
-    pub async fn sync_tunnel_pool(&self, pool_id: u64) -> Result<()> {
-        let pool = self.db.get_tunnel_pool(pool_id).await?;
-        let router = crate::router::get_router(&self.db, pool.router_id)
-            .await
-            .map_err(|e| anyhow!("failed to load router {}: {}", pool.router_id, e))?;
-        let tr = router.tunnel().context("router does not support tunnels")?;
-
-        let private_key = pool.private_key.as_str().to_string();
-        // The stored pair has to agree with itself before it is pushed: a
-        // public key that is not this private key's would be handed to every
-        // node and none of them could hand shake.
-        let derived = lnvps_api_common::wireguard_public_key(&private_key)?;
-        if derived != pool.public_key {
-            bail!(
-                "Tunnel pool {} has a public key that its private key does not produce; \
-                 refusing to configure an interface nobody could connect to",
-                pool.id
-            );
-        }
-
-        // Named from the pool's id under a fixed prefix, so a managed
-        // interface can never be confused with one the operator of the route
-        // server configured themselves.
-        let interface = pool.interface();
-
-        let existing = tr
-            .list_tunnels()
-            .await
-            .map_err(|e| anyhow!("failed to list tunnels: {}", e))?
-            .into_iter()
-            .find(|t| t.name == interface);
-
-        let desired = crate::router::Tunnel {
-            id: existing.as_ref().and_then(|t| t.id.clone()),
-            name: interface.clone(),
-            // The address the data plane listens on. Recorded so the interface
-            // and the endpoint peers are told to dial cannot disagree.
-            local_addr: Some(pool.listen_addr.clone()),
-            remote_addr: None,
-            enabled: pool.enabled,
-            config: crate::router::TunnelConfig::Wireguard(crate::router::WireguardConfig {
-                listen_port: Some(pool.listen_port),
-                private_key: Some(private_key),
-                public_key: Some(lnvps_api_common::wireguard_key_to_base64(&pool.public_key)),
-                // Peers are pushed per allocation, not here. Sending an empty
-                // list would be read as "this interface has no peers".
-                peers: vec![],
-            }),
-        };
-
-        match &existing {
-            None => {
-                info!(
-                    "Creating WireGuard interface {} on router {}",
-                    interface, pool.router_id
-                );
-                tr.add_tunnel(&desired)
-                    .await
-                    .map_err(|e| anyhow!("failed to create tunnel interface: {}", e))?;
-            }
-            Some(current) => {
-                // Re-applying recreates the interface on the Linux backend,
-                // which drops every peer with it. So it is only done when the
-                // interface is actually wrong — a node whose tunnel is working
-                // must not be cut because a pool was renamed.
-                let current_key = match &current.config {
-                    crate::router::TunnelConfig::Wireguard(c) => c.public_key.clone(),
-                    _ => None,
-                };
-                let current_port = match &current.config {
-                    crate::router::TunnelConfig::Wireguard(c) => c.listen_port,
-                    _ => None,
-                };
-                let want_key = lnvps_api_common::wireguard_key_to_base64(&pool.public_key);
-                let key_drifted = current_key.as_deref() != Some(want_key.as_str());
-                let port_drifted = current_port != Some(pool.listen_port);
-
-                if key_drifted || port_drifted {
-                    warn!(
-                        "WireGuard interface {} on router {} has drifted (key_changed={}, \
-                         port_changed={}); re-applying, which drops its peers until they are \
-                         pushed again",
-                        interface, pool.router_id, key_drifted, port_drifted
-                    );
-                    tr.update_tunnel(&desired)
-                        .await
-                        .map_err(|e| anyhow!("failed to update tunnel interface: {}", e))?;
-                } else if current.enabled != pool.enabled {
-                    let id = current.id.as_deref().unwrap_or(interface.as_str());
-                    tr.set_tunnel_enabled(id, pool.enabled)
-                        .await
-                        .map_err(|e| anyhow!("failed to toggle tunnel interface: {}", e))?;
-                }
-            }
-        }
-
-        // Refresh the observed-state cache so the admin API stops showing the
-        // interface as missing the moment it exists.
-        if let Ok(tunnels) = tr.list_tunnels().await {
-            for t in &tunnels {
-                if let Err(e) = self.db.upsert_router_tunnel(&t.to_db(pool.router_id)).await {
-                    warn!("Failed to refresh tunnel cache: {}", e);
-                }
-            }
-        }
-
-        // Whatever happened above, the interface now has to carry the peers
-        // that were allocated from this pool. This matters most in the case the
-        // push above just created or re-applied it: on Linux that is a fresh
-        // interface with no peers at all, and every node on it is cut until
-        // they are put back.
-        self.reconcile_tunnel_peers(pool.id).await?;
-        Ok(())
-    }
-
-    /// Reconcile the peers, addresses and routes on a pool's interface against
-    /// the tunnels allocated from it.
-    ///
-    /// The `tunnel` table is the desired state and the router is the observed
-    /// one, exactly as with host state. A peer that has vanished from a route
-    /// server is drift to put back and report, not an allocation to forget:
-    /// forgetting it would hand the node's addresses to somebody else while the
-    /// node still believes they are its own.
-    ///
-    /// Returns what had drifted, so a caller running this on a schedule can say
-    /// whether anything was wrong rather than only that it ran.
-    pub async fn reconcile_tunnel_peers(&self, pool_id: u64) -> Result<TunnelPeerDrift> {
-        let pool = self.db.get_tunnel_pool(pool_id).await?;
-        let router = crate::router::get_router(&self.db, pool.router_id)
-            .await
-            .map_err(|e| anyhow!("failed to load router {}: {}", pool.router_id, e))?;
-        let tr = router.tunnel().context("router does not support tunnels")?;
-        let interface = pool.interface();
-
-        let observed = tr
-            .list_tunnels()
-            .await
-            .map_err(|e| anyhow!("failed to list tunnels: {}", e))?
-            .into_iter()
-            .find(|t| t.name == interface);
-        // Peers are configured *on* an interface, so there is nothing to
-        // reconcile against until it exists. Creating it here would duplicate
-        // `sync_tunnel_pool` and hide the fact that it never ran.
-        let Some(observed) = observed else {
-            bail!(
-                "Tunnel pool {pool_id}'s interface {interface} is not configured on router {}; \
-                 run SyncTunnelPool first",
-                pool.router_id
-            );
-        };
-        let observed_peers = match &observed.config {
-            crate::router::TunnelConfig::Wireguard(c) => c.peers.clone(),
-            _ => bail!("Tunnel pool {pool_id}'s interface {interface} is not a WireGuard tunnel"),
-        };
-
-        // What is behind each node peer is recomputed from the guest
-        // assignments before the plan is built, so the planner can read it
-        // without knowing that marketplace nodes exist.
-        crate::provisioner::refresh_node_routes(&self.db, &pool).await?;
-        let plan = crate::provisioner::plan_pool(&self.db, &pool).await?;
-        let mut drift = TunnelPeerDrift::default();
-
-        for want in &plan.peers {
-            match observed_peers
-                .iter()
-                .find(|p| p.public_key == want.public_key)
-            {
-                // Allowed IPs are compared as a set: `wg` reports them in its
-                // own order, and re-pushing on every reconcile because of that
-                // would rewrite the anti-spoof list of a working peer forever.
-                Some(have) if same_allowed_ips(have, want) => continue,
-                Some(_) => drift.changed.push(want.public_key.clone()),
-                None => drift.missing.push(want.public_key.clone()),
-            }
-            tr.set_tunnel_peer(&interface, want)
-                .await
-                .map_err(|e| anyhow!("failed to configure peer on {interface}: {}", e))?;
-        }
-
-        for have in &observed_peers {
-            if plan.peers.iter().any(|p| p.public_key == have.public_key) {
-                continue;
-            }
-            // LNVPS owns `wgln*` interfaces outright, so a peer no tunnel
-            // claims is either a revoked allocation that was never cleaned up
-            // or somebody else's key on our route server. Both are removed.
-            drift.unclaimed.push(have.public_key.clone());
-            tr.remove_tunnel_peer(&interface, &have.public_key)
-                .await
-                .map_err(|e| anyhow!("failed to remove peer from {interface}: {}", e))?;
-        }
-
-        tr.sync_tunnel_addresses(&interface, &plan.addresses)
-            .await
-            .map_err(|e| anyhow!("failed to configure addresses on {interface}: {}", e))?;
-        tr.sync_tunnel_routes(&interface, &plan.routes)
-            .await
-            .map_err(|e| anyhow!("failed to configure routes on {interface}: {}", e))?;
-
-        if !drift.is_empty() {
-            warn!(
-                "Tunnel pool {pool_id} on router {} had drifted: {drift}",
-                pool.router_id
-            );
-        }
-        Ok(drift)
-    }
-
-    /// Push one node's peer onto its route server.
-    ///
-    /// Used when a single allocation changes — a node asking for its tunnel, a
-    /// guest getting an address — so it does not wait behind a reconcile of
-    /// every other node on the same route server.
-    pub async fn sync_node_tunnel(&self, tunnel_id: u64) -> Result<()> {
-        let tunnel = self.db.get_tunnel(tunnel_id).await?;
-        let pool_id = tunnel.pool_id.ok_or_else(|| {
-            anyhow!("Tunnel {tunnel_id} was not allocated from a pool, so there is no interface")
-        })?;
-        let pool = self.db.get_tunnel_pool(pool_id).await?;
-        let router = crate::router::get_router(&self.db, pool.router_id)
-            .await
-            .map_err(|e| anyhow!("failed to load router {}: {}", pool.router_id, e))?;
-        let tr = router.tunnel().context("router does not support tunnels")?;
-        let interface = pool.interface();
-
-        // The peer's own share of the pool plan, rather than a second
-        // calculation of what one tunnel needs: the addresses and routes are
-        // per-interface, so one node's change is applied by re-stating the
-        // whole interface's addressing, and only its own peer is pushed.
-        crate::provisioner::refresh_node_routes(&self.db, &pool).await?;
-        let plan = crate::provisioner::plan_pool(&self.db, &pool).await?;
-        let key = tunnel
-            .peer_pubkey
-            .as_deref()
-            .map(lnvps_api_common::wireguard_key_to_base64);
-
-        match key
-            .as_ref()
-            .and_then(|k| plan.peers.iter().find(|p| &p.public_key == k))
-        {
-            Some(peer) => tr
-                .set_tunnel_peer(&interface, peer)
-                .await
-                .map_err(|e| anyhow!("failed to configure peer on {interface}: {}", e))?,
-            // A tunnel that is disabled or has never presented a key has no
-            // peer to push. Removing whatever is there under its key is the
-            // same statement in the other direction.
-            None => {
-                if let Some(key) = &key {
-                    tr.remove_tunnel_peer(&interface, key)
-                        .await
-                        .map_err(|e| anyhow!("failed to remove peer from {interface}: {}", e))?;
-                }
-            }
-        }
-
-        tr.sync_tunnel_addresses(&interface, &plan.addresses)
-            .await
-            .map_err(|e| anyhow!("failed to configure addresses on {interface}: {}", e))?;
-        tr.sync_tunnel_routes(&interface, &plan.routes)
-            .await
-            .map_err(|e| anyhow!("failed to configure routes on {interface}: {}", e))?;
         Ok(())
     }
 
@@ -3182,10 +2838,12 @@ impl Worker {
                 name,
                 enabled,
             } => {
-                self.toggle_tunnel(*router_id, name, *enabled).await?;
+                self.tunnels()
+                    .toggle_tunnel(*router_id, name, *enabled)
+                    .await?;
             }
             WorkJob::SyncTunnelPool { pool_id } => {
-                self.sync_tunnel_pool(*pool_id).await?;
+                self.tunnels().sync_tunnel_pool(*pool_id).await?;
             }
             WorkJob::RemoveTunnelInterface {
                 router_id,
@@ -3194,10 +2852,10 @@ impl Worker {
                 self.remove_tunnel_interface(*router_id, interface).await?;
             }
             WorkJob::ReconcileTunnelPeers { pool_id } => {
-                self.reconcile_tunnel_peers(*pool_id).await?;
+                self.tunnels().reconcile_tunnel_peers(*pool_id).await?;
             }
             WorkJob::SyncNodeTunnel { tunnel_id } => {
-                self.sync_node_tunnel(*tunnel_id).await?;
+                self.tunnels().sync_node_tunnel(*tunnel_id).await?;
             }
             WorkJob::DeleteVm {
                 vm_id,
@@ -5822,7 +5480,7 @@ mod tests {
     async fn test_sync_router_state() -> Result<()> {
         use crate::mocks::MockRouter;
         use crate::router::{
-            BgpPeerDirection, BgpSession, GreConfig, Router as _, Tunnel, TunnelConfig,
+            BgpPeerDirection, BgpSession, GreConfig, ObservedInterface, Router as _, TunnelConfig,
         };
         use lnvps_db::{Router, RouterKind};
 
@@ -5847,7 +5505,7 @@ mod tests {
         mr.clear().await;
         mr.tunnel()
             .unwrap()
-            .add_tunnel(&Tunnel {
+            .add_tunnel(&ObservedInterface {
                 id: None,
                 name: "gre1".to_string(),
                 local_addr: None,
@@ -5999,7 +5657,7 @@ mod tests {
     #[tokio::test]
     async fn test_toggle_tunnel() -> Result<()> {
         use crate::mocks::MockRouter;
-        use crate::router::{GreConfig, Router as _, Tunnel, TunnelConfig};
+        use crate::router::{GreConfig, ObservedInterface, Router as _, TunnelConfig};
         use lnvps_db::{Router, RouterKind};
 
         let db = Arc::new(MockDb::empty());
@@ -6021,7 +5679,7 @@ mod tests {
         mr.clear().await;
         mr.tunnel()
             .unwrap()
-            .add_tunnel(&Tunnel {
+            .add_tunnel(&ObservedInterface {
                 id: None,
                 name: "gre1".to_string(),
                 local_addr: None,
@@ -6033,7 +5691,7 @@ mod tests {
             .unwrap();
 
         let worker = setup_worker(db.clone()).await?;
-        worker.toggle_tunnel(1, "gre1", false).await?;
+        worker.tunnels().toggle_tunnel(1, "gre1", false).await?;
 
         // The cached tunnel should reflect the disabled state after refresh.
         let tunnels = db.list_router_tunnels(1).await?;
@@ -6165,7 +5823,7 @@ mod tests {
         mr.clear().await;
 
         let worker = setup_worker(db.clone()).await?;
-        worker.sync_tunnel_pool(pool_id).await?;
+        worker.tunnels().sync_tunnel_pool(pool_id).await?;
 
         let interface = format!("wgln{pool_id}");
         let peers = mr.peers(&interface).await;
@@ -6216,19 +5874,19 @@ mod tests {
         let mr = MockRouter::new();
         mr.clear().await;
         let worker = setup_worker(db.clone()).await?;
-        worker.sync_tunnel_pool(pool_id).await?;
+        worker.tunnels().sync_tunnel_pool(pool_id).await?;
         let interface = format!("wgln{pool_id}");
 
         // Nothing changed: a working peer must not be rewritten on every poll,
         // and `wg` reports allowed IPs in its own order.
-        let drift = worker.reconcile_tunnel_peers(pool_id).await?;
+        let drift = worker.tunnels().reconcile_tunnel_peers(pool_id).await?;
         assert!(drift.is_empty(), "{drift}");
 
         // The route server lost the peer (a reboot without persistence).
         let tr = mr.tunnel().unwrap();
         let key = lnvps_api_common::wireguard_key_to_base64(&[0x11u8; 32]);
         tr.remove_tunnel_peer(&interface, &key).await.unwrap();
-        let drift = worker.reconcile_tunnel_peers(pool_id).await?;
+        let drift = worker.tunnels().reconcile_tunnel_peers(pool_id).await?;
         assert_eq!(drift.missing, vec![key.clone()]);
         assert_eq!(mr.peers(&interface).await.len(), 1, "not put back");
 
@@ -6244,7 +5902,7 @@ mod tests {
         )
         .await
         .unwrap();
-        let drift = worker.reconcile_tunnel_peers(pool_id).await?;
+        let drift = worker.tunnels().reconcile_tunnel_peers(pool_id).await?;
         assert_eq!(drift.changed, vec![key.clone()]);
         assert_eq!(
             mr.peers(&interface).await[0].allowed_ips,
@@ -6262,7 +5920,7 @@ mod tests {
         )
         .await
         .unwrap();
-        let drift = worker.reconcile_tunnel_peers(pool_id).await?;
+        let drift = worker.tunnels().reconcile_tunnel_peers(pool_id).await?;
         assert_eq!(drift.unclaimed, vec!["c3RyYXk=".to_string()]);
         assert_eq!(mr.peers(&interface).await.len(), 1);
 
@@ -6283,6 +5941,7 @@ mod tests {
         let worker = setup_worker(db.clone()).await?;
 
         let err = worker
+            .tunnels()
             .reconcile_tunnel_peers(pool_id)
             .await
             .expect_err("peers were reconciled onto an interface that is not there");
@@ -6308,13 +5967,13 @@ mod tests {
         let worker = setup_worker(db.clone()).await?;
         // The interface exists but has no peers, which is exactly the state
         // right after a node asks for its tunnel.
-        worker.sync_tunnel_pool(pool_id).await?;
+        worker.tunnels().sync_tunnel_pool(pool_id).await?;
         let interface = format!("wgln{pool_id}");
         let tr = mr.tunnel().unwrap();
         let key = lnvps_api_common::wireguard_key_to_base64(&[0x11u8; 32]);
         tr.remove_tunnel_peer(&interface, &key).await.unwrap();
 
-        worker.sync_node_tunnel(tunnel.id).await?;
+        worker.tunnels().sync_node_tunnel(tunnel.id).await?;
         assert_eq!(mr.peers(&interface).await.len(), 1);
 
         // Disabling the allocation is a statement in the other direction: the
@@ -6325,7 +5984,7 @@ mod tests {
             ..tunnel.clone()
         })
         .await?;
-        worker.sync_node_tunnel(tunnel.id).await?;
+        worker.tunnels().sync_node_tunnel(tunnel.id).await?;
         assert!(mr.peers(&interface).await.is_empty());
 
         mr.clear().await;
@@ -6350,6 +6009,7 @@ mod tests {
 
         let worker = setup_worker(db.clone()).await?;
         let err = worker
+            .tunnels()
             .sync_node_tunnel(tunnel_id)
             .await
             .expect_err("a pool-less tunnel was pushed to an interface");
@@ -6373,7 +6033,7 @@ mod tests {
         mr.clear().await;
 
         let worker = setup_worker(db.clone()).await?;
-        worker.sync_tunnel_pool(pool_id).await?;
+        worker.tunnels().sync_tunnel_pool(pool_id).await?;
 
         let tunnels = mr.tunnel().unwrap().list_tunnels().await.unwrap();
         assert_eq!(tunnels.len(), 1);
@@ -6401,7 +6061,7 @@ mod tests {
         // Syncing again is a no-op rather than a recreate: re-applying drops
         // every peer with the interface, so a working node must not be cut by
         // a routine sync.
-        worker.sync_tunnel_pool(pool_id).await?;
+        worker.tunnels().sync_tunnel_pool(pool_id).await?;
         let after = mr.tunnel().unwrap().list_tunnels().await.unwrap();
         assert_eq!(after.len(), 1);
         assert_eq!(after[0].id, tunnels[0].id);
@@ -6428,7 +6088,7 @@ mod tests {
         mr.clear().await;
         let worker = setup_worker(db.clone()).await?;
 
-        assert!(worker.sync_tunnel_pool(pool_id).await.is_err());
+        assert!(worker.tunnels().sync_tunnel_pool(pool_id).await.is_err());
         assert!(
             mr.tunnel()
                 .unwrap()
@@ -6456,7 +6116,7 @@ mod tests {
         mr.clear().await;
 
         let worker = setup_worker(db.clone()).await?;
-        worker.sync_tunnel_pool(pool_id).await?;
+        worker.tunnels().sync_tunnel_pool(pool_id).await?;
         assert_eq!(mr.tunnel().unwrap().list_tunnels().await.unwrap().len(), 1);
 
         let interface = db.get_tunnel_pool(pool_id).await?.interface();
